Cambio en las consultas

+++ src/main/java/com/cuenta/bancaria/cuenta/bancaria/controller/CuentaController.java
@@ -12,6 +12,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.ObjectUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -53,15 +54,20 @@ public class CuentaController {
 	@PostMapping
 	public ResponseEntity<?> create(@Validated @RequestBody CuentaEntradaDto cuentaEntradaDto) {
 		try {
-			Optional<Persona> personaEncontrada = personaService
-					.obtenerPorIdentificacion(cuentaEntradaDto.getIdentificacion());
-			if (!personaEncontrada.isPresent()) {
-				return new ResponseEntity<>("No existe la persona con esta identificación", HttpStatus.BAD_REQUEST);
+			if (!ObjectUtils.isEmpty(cuentaEntradaDto.getIdentificacion())) {
+				Optional<Persona> personaEncontrada = personaService
+						.obtenerPorIdentificacion(cuentaEntradaDto.getIdentificacion());
+				if (personaEncontrada.isPresent()) {
+					cuentaEntradaDto.setIdPersona(personaEncontrada.get().getIdPersona());
+				} else {
+					return new ResponseEntity<>("No existe la persona con esta identificación", HttpStatus.BAD_REQUEST);
+				}
 			}
-			Optional<Cliente> cliente = clienteService.obtenerPorIdCliente(personaEncontrada.get().getIdPersona());
+
+			Optional<Cliente> cliente = clienteService.obtenerPorIdCliente(cuentaEntradaDto.getIdPersona());
 			if (cliente.isPresent()) {
 				if (EstadoEmun.INACTIVO.getDescripcion().equals(cliente.get().getEstado())) {
-
+					return new ResponseEntity<>("El cliente se encuentra inactivo", HttpStatus.BAD_REQUEST);
 				} else {
 					Cuenta cuenta = new Cuenta();
 					cuenta.setNumero(Integer.parseInt(cuentaEntradaDto.getNumero()));
@@ -82,16 +88,38 @@ public class CuentaController {
 	}
 
 	@GetMapping
-	public ResponseEntity<List<Cuenta>> read() {
-		List<Cuenta> list = service.read();
-		return new ResponseEntity<List<Cuenta>>(list, HttpStatus.OK);
+	public ResponseEntity<?> obtenerCuentaPorCliente(@Validated @RequestBody CuentaEntradaDto cuentaEntradaDto) {
+		try {
+			List<Cuenta> listaCuenta = service.obtenerPorCliente(cuentaEntradaDto.getIdPersona());
+			if (null == listaCuenta || listaCuenta.isEmpty()) {
+				return new ResponseEntity<>("No existe cuenta con el id", HttpStatus.OK);
+			} else {
+				return new ResponseEntity<List<Cuenta>>(listaCuenta, HttpStatus.OK);
+			}
+
+		} catch (Exception e) {
+			log.error("Por favor comuniquese con el administrador", e);
+			return new ResponseEntity<>("Por favor comuniquese con el administrador", HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+
 	}
 
 	@PutMapping
-	public ResponseEntity<?> update(@Validated @RequestBody Cuenta cuenta) {
+	public ResponseEntity<?> update(@Validated @RequestBody CuentaEntradaDto cuentaEntradaDto) {
 		try {
-			Cuenta cuentaGuardada = service.update(cuenta);
-			return new ResponseEntity<Cuenta>(cuentaGuardada, HttpStatus.OK);
+			Optional<Cuenta> cuenta = service.obtenerPorId(cuentaEntradaDto.getIdCuenta());
+			if (cuenta.isPresent()) {
+				Cuenta cuentaActualizar = cuenta.get();
+				cuentaActualizar.setNumero(Integer.parseInt(cuentaEntradaDto.getNumero()));
+				cuentaActualizar.setTipoCuenta(cuentaEntradaDto.getTipoCuenta());
+				cuentaActualizar
+						.setSaldoInicial(BigDecimal.valueOf(Double.valueOf(cuentaEntradaDto.getSaldoInicial())));
+				cuentaActualizar.setEstado(cuentaEntradaDto.getEstado());
+				Cuenta cuentaGuardada = service.update(cuentaActualizar);
+				return new ResponseEntity<Cuenta>(cuentaGuardada, HttpStatus.OK);
+			}
+			return new ResponseEntity<>("No se encuentra la cuenta", HttpStatus.BAD_REQUEST);
+
 		} catch (Exception e) {
 			log.error("Por favor comuniquese con el administrador", e);
 			return new ResponseEntity<>("Por favor comuniquese con el administrador", HttpStatus.INTERNAL_SERVER_ERROR);
@@ -101,8 +129,14 @@ public class CuentaController {
 	@DeleteMapping("/{id}")
 	public ResponseEntity<?> delete(@PathVariable("id") Long id) {
 		try {
-			service.delete(id);
-			return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
+			Optional<Cuenta> cuenta = service.obtenerPorId(id);
+			if (cuenta.isPresent()) {
+				service.delete(id);
+				return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
+			}
+			return new ResponseEntity<String>("No se puede eliminar la cuenta con id: " + id + " no existe el registro",
+					HttpStatus.BAD_REQUEST);
+
 		} catch (Exception e) {
 			log.error("Por favor comuniquese con el administrador", e);
 			return new ResponseEntity<>("Por favor comuniquese con el administrador", HttpStatus.INTERNAL_SERVER_ERROR);
